+++ src/models/node_hardware_node.rs
@@ -98,7 +98,7 @@ pub struct NodeHardwareNode {
     pub product: Option<String>,
     /// Size of RAM in bytes.
     #[serde(rename = "ram")]
-    pub ram: Option<i32>,
+    pub ram: Option<u64>,
     /// Serial number of this node.
     #[serde(rename = "serial_number")]
     pub serial_number: Option<String>,
