+++ src/models/result_directories_usage_data_item.rs
@@ -20,10 +20,10 @@ pub struct ResultDirectoriesUsageDataItem {
     pub lin: i32,
     /// Logical size directory in bytes.
     #[serde(rename = "log_size_sum")]
-    pub log_size_sum: i32,
+    pub log_size_sum: u64,
     /// Logical size sum of overflow in bytes.
     #[serde(rename = "log_size_sum_overflow")]
-    pub log_size_sum_overflow: i32,
+    pub log_size_sum_overflow: u64,
     /// Name of directory.
     #[serde(rename = "name")]
     pub name: String,
@@ -35,5 +35,5 @@ pub struct ResultDirectoriesUsageDataItem {
     pub parent: i32,
     /// Physical size directory in bytes.
     #[serde(rename = "phys_size_sum")]
-    pub phys_size_sum: i32,
+    pub phys_size_sum: u64,
 }
